+++ database/migrations/2025_11_29_101437_drop_mitigasi_id_from_penilaian_table.php
@@ -4,18 +4,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::table('penilaian', function (Blueprint $table) {
-            // hapus dulu foreign key
-            $table->dropForeign(['mitigasi_id']);
 
-            // baru hapus kolom
-            $table->dropColumn('mitigasi_id');
+            if (Schema::hasColumn('penilaian', 'mitigasi_id')) {
+                $table->dropForeign(['mitigasi_id']);
+                $table->dropColumn('mitigasi_id');
+            }
+
         });
     }
+
 
     public function down(): void
     {
